+++ quimb/operator/__init__.py
@@ -1,0 +1,28 @@
+"""Tools for 'symbolically' and consitently defining operators and
+hamiltonians, which can thenvbe built out into various matrix or tensor network
+forms such as sparse matrices and MPOs.
+"""
+
+from .builder import (
+    SparseOperatorBuilder,
+    get_mat,
+)
+from .hilbertspace import (
+    HilbertSpace,
+)
+from .models import (
+    fermi_hubbard_from_edges,
+    fermi_hubbard_spinless_from_edges,
+    heisenberg_from_edges,
+    rand_operator,
+)
+
+__all__ = (
+    "fermi_hubbard_from_edges",
+    "fermi_hubbard_spinless_from_edges",
+    "get_mat",
+    "heisenberg_from_edges",
+    "HilbertSpace",
+    "rand_operator",
+    "SparseOperatorBuilder",
+)